requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
nbformat
nbconvert